improve we toolbar
